% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_clinical.R
\name{adverse_events}
\alias{adverse_events}
\title{Adverse events}
\usage{
adverse_events(db = NULL, set.name = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{set.name}{Optional clinical observation set name}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
One row per adverse event
}
\description{
Adverse events
}
