% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{args}
\alias{args}
\title{Query inputs and their values}
\usage{
args(...)
}
\arguments{
\item{...}{Bindings}
}
\value{
Query section
}
\description{
Each argument has the form \code{?variable <- value}. The value is
evaluated in the calling environment. A value of length greater than one
(or wrapped in \code{I()}) is bound as a collection (\code{[?variable ...]}).
}
\examples{
args(?sample-id <- "H37001-001", ?gene <- c("TP53", "KRAS"))
}
