% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{c_query}
\alias{c_query}
\title{Combine queries}
\usage{
c_query(a, b)
}
\arguments{
\item{a, b}{Queries}
}
\value{
A \code{patternq_query}
}
\description{
Concatenates the sections and arguments of two queries.
}
