% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_clinical.R
\name{clinical_observation_sets}
\alias{clinical_observation_sets}
\title{Clinical observation sets}
\usage{
clinical_observation_sets(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
\code{clinical_observation_set_name}, \code{..._description}
}
\description{
Clinical observation sets
}
