% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_expression.R
\name{compare_samples}
\alias{compare_samples}
\title{Compare expression between two samples}
\usage{
compare_samples(
  sample.a,
  sample.b,
  db = NULL,
  db.b = db,
  measurement = "tpm",
  min.avg = 0.5,
  pseudocount = 1
)
}
\arguments{
\item{sample.a}{Reference sample (e.g. baseline)}

\item{sample.b}{Comparison sample (e.g. later timepoint)}

\item{db}{Database (both samples); use \code{db.b} if \code{sample.b}
lives elsewhere}

\item{db.b}{Database of \code{sample.b}}

\item{measurement}{Measurement attribute}

\item{min.avg}{Drop genes with \code{avg_log10} below this (low expression)}

\item{pseudocount}{See \code{\link{log_fold_change}}}
}
\value{
\code{\link{log_fold_change}} output, ordered by \code{lfc}
}
\description{
Compare expression between two samples
}
