% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{current_db}
\alias{current_db}
\title{The default database for this session, if set}
\usage{
current_db()
}
\value{
The database name or \code{NULL}
}
\description{
The default database for this session, if set
}
