% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{d}
\alias{d}
\title{Convert R expressions to Datalog terms}
\usage{
d(...)
}
\arguments{
\item{...}{Expressions forming one clause}
}
\value{
A list representing the clause
}
\description{
Builds one Datalog clause from R expressions (see the package DSL notes in
\code{\link{query}}).
}
\examples{
d(?m, measurement/sample, ?s)
}
