% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference.R
\name{dataset_info}
\alias{dataset_info}
\title{Dataset entity metadata}
\usage{
dataset_info(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
One-row data frame
}
\description{
The dataset entity stored in the database: name, description, doi, url.
}
