% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{dataset_summary}
\alias{dataset_summary}
\alias{dataset_summary_query}
\title{Assays and measurement sets in a dataset}
\usage{
dataset_summary(db = NULL, ...)

dataset_summary_query()
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
One row per measurement set: \code{assay_name},
  \code{assay_technology}, \code{measurement_set_name}
}
\description{
Assays and measurement sets in a dataset
}
