% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{datoms}
\alias{datoms}
\title{Read raw datoms from an index}
\usage{
datoms(index, components = list(), db = NULL, offset = 0, limit = 1000)
}
\arguments{
\item{index}{One of \code{"eavt"}, \code{"aevt"}, \code{"avet"}, \code{"vaet"}}

\item{components}{Index components, e.g. \code{list(":sample/id")}}

\item{db}{Database name}

\item{offset, limit}{Paging}
}
\value{
A \code{data.frame} with columns \code{e}, \code{a}, \code{v}, \code{tx}
}
\description{
Read raw datoms from an index
}
