% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference.R
\name{db_idents}
\alias{db_idents}
\title{All idents in a database}
\usage{
db_idents(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
\code{db_id}, \code{db_ident}
}
\description{
All idents in a database
}
