% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{find}
\alias{find}
\title{The find portion of a query}
\usage{
find(...)
}
\arguments{
\item{...}{Variables, aggregates (e.g. \code{count(?s)}) or \code{\link{pull}}
expressions}
}
\value{
Query section
}
\description{
The find portion of a query
}
