% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference.R
\name{gene_coordinates}
\alias{gene_coordinates}
\alias{gene_coordinates_query}
\title{Genes with genomic coordinates}
\usage{
gene_coordinates(db = NULL, genes = NULL, ...)

gene_coordinates_query(genes = NULL)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{genes}{Optional HGNC symbols}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
\code{hgnc_symbol}, \code{assembly}, \code{contig}, \code{strand},
  \code{start}, \code{end}
}
\description{
Genes with genomic coordinates
}
