% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference.R
\name{gene_products}
\alias{gene_products}
\alias{gene_products_query}
\title{Gene products and their genes}
\usage{
gene_products(db = NULL, ...)

gene_products_query()
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
\code{gene_product_id}, \code{hgnc_symbol}
}
\description{
Gene products and their genes
}
