% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference.R
\name{gene_symbols}
\alias{gene_symbols}
\alias{gene_symbols_query}
\title{HGNC gene symbols}
\usage{
gene_symbols(db = NULL, ...)

gene_symbols_query()
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
Character vector
}
\description{
HGNC gene symbols
}
