% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{generate_or}
\alias{generate_or}
\title{Generate an or clause over alternative values}
\usage{
generate_or(alternatives, ...)
}
\arguments{
\item{alternatives}{Values}

\item{...}{Partial clause}
}
\value{
Clause
}
\description{
\code{generate_or(c("A", "B"), ?p, epitope/id)} is equivalent to
\code{or(d(?p, epitope/id, "A"), d(?p, epitope/id, "B"))}.
}
