% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{ident_name}
\alias{ident_name}
\title{Keyword name without namespace}
\usage{
ident_name(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector
}
\description{
\code{":variant.impact/high"} becomes \code{"high"}. Non-keyword values are
returned unchanged.
}
