% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{is_pull_query}
\alias{is_pull_query}
\title{Does a query's find spec contain a pull expression?}
\usage{
is_pull_query(q)
}
\arguments{
\item{q}{A query}
}
\value{
Logical
}
\description{
Does a query's find spec contain a pull expression?
}
