% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{kaplan_meier}
\alias{kaplan_meier}
\title{Kaplan-Meier estimate}
\usage{
kaplan_meier(time, event)
}
\arguments{
\item{time}{Numeric times}

\item{event}{Logical, \code{TRUE} = event, \code{FALSE} = censored}
}
\value{
\code{data.frame} of \code{time}, \code{n.risk}, \code{n.event},
  \code{n.censor}, \code{surv}
}
\description{
Kaplan-Meier estimate
}
