% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets available to your API key}
\usage{
list_datasets()
}
\value{
A \code{data.frame} with one row per dataset: \code{dataset}
  (dataset name), \code{db} (current database name), \code{patient_count},
  \code{sample_count}, \code{assays} and \code{tags} (list columns)
}
\description{
List datasets available to your API key
}
