% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_measurements.R
\name{measurement_matrix_by_name}
\alias{measurement_matrix_by_name}
\title{Download a measurement matrix by name}
\usage{
measurement_matrix_by_name(matrix.name, db = NULL)
}
\arguments{
\item{matrix.name}{Measurement matrix name (see \code{\link{measurement_matrices}})}

\item{db}{Database name (defaults to \code{\link{current_db}})}
}
\value{
\code{data.frame}
}
\description{
Download a measurement matrix by name
}
