% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_clinical.R
\name{measurement_set_entities}
\alias{measurement_set_entities}
\alias{cell_populations}
\alias{tcrs}
\alias{otus}
\alias{sgbs}
\title{Entities of a measurement set}
\usage{
cell_populations(measurement.set, db = NULL, ...)

tcrs(measurement.set, db = NULL, ...)

otus(measurement.set, db = NULL, ...)

sgbs(measurement.set, db = NULL, ...)
}
\arguments{
\item{measurement.set}{Measurement set name}

\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
One row per entity
}
\description{
Cell populations, TCRs, OTUs and SGBs (metagenomic species) attached to a
measurement set.
}
