% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_measurements.R
\name{measurement_sets}
\alias{measurement_sets}
\title{Measurement sets with their assay and size}
\usage{
measurement_sets(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
\code{assay_name}, \code{assay_technology},
  \code{measurement_set_name}, \code{measurement_count}
}
\description{
Measurement sets with their assay and size
}
