% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{not_join}
\alias{not_join}
\title{not-join clause}
\usage{
not_join(vars, ...)
}
\arguments{
\item{vars}{Variables to join on, built with \code{\link{d}}}

\item{...}{Clauses}
}
\value{
Clause
}
\description{
not-join clause
}
