% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_expression.R
\name{percentile_rank}
\alias{percentile_rank}
\title{Percentile rank of a value within a distribution}
\usage{
percentile_rank(values, x)
}
\arguments{
\item{values}{Reference values}

\item{x}{Value(s) to rank}
}
\value{
Percentile(s) in [0, 100]
}
\description{
Mid-rank percentile: (count below + half the ties) / n, as a percentage.
}
