% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_expression.R
\name{plot_fold_change}
\alias{plot_fold_change}
\title{Fold change bars for selected genes}
\usage{
plot_fold_change(change, genes = NULL, title = "Change in gene expression")
}
\arguments{
\item{change}{Output of \code{\link{compare_samples}}}

\item{genes}{Genes to show (default: top 15 up and 15 down)}

\item{title}{Plot title}
}
\value{
A plotly object
}
\description{
Fold change bars for selected genes
}
