% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gene_expression}
\alias{plot_gene_expression}
\title{Gene expression bar plot}
\usage{
plot_gene_expression(
  expr,
  title = "Gene expression",
  ylab = "value",
  log = FALSE
)
}
\arguments{
\item{expr}{A data frame from \code{\link{gene_expression}} (needs
\code{sample_id}, \code{hgnc_symbol}, \code{value})}

\item{title}{Plot title}

\item{ylab}{Y axis label}

\item{log}{Use a log scale y axis}
}
\value{
A plotly object
}
\description{
Gene expression bar plot
}
