% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{query}
\alias{query}
\title{Construct a Datalog query}
\usage{
query(...)
}
\arguments{
\item{...}{Query sections}
}
\value{
A \code{patternq_query}
}
\description{
Combines \code{\link{find}}, \code{\link{where}} and optionally
\code{\link{args}} and \code{\link{with_vars}} sections into a
\code{patternq_query}.
}
