% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{query_json}
\alias{query_json}
\title{Serialize a query to the JSON sent to the query service}
\usage{
query_json(q, pretty = FALSE, timeout = 30, refresh.cache = FALSE)
}
\arguments{
\item{q}{A \code{patternq_query}}

\item{pretty}{Pretty-print}

\item{timeout}{Query timeout in seconds}

\item{refresh.cache}{Ask the service to recompute and re-cache the result}
}
\value{
JSON string
}
\description{
Serialize a query to the JSON sent to the query service
}
