% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference.R
\name{reference_names}
\alias{reference_names}
\alias{gdc_anatomic_sites}
\alias{proteins}
\alias{epitopes}
\alias{cell_types}
\alias{meddra_diseases}
\alias{drugs}
\title{Reference vocabularies}
\usage{
gdc_anatomic_sites(db = NULL, ...)

proteins(db = NULL, ...)

epitopes(db = NULL, ...)

cell_types(db = NULL, ...)

meddra_diseases(db = NULL, ...)

drugs(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
Character vector
}
\description{
Names of reference entities present in a database.
}
