% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backpressure.R
\name{retry_policy}
\alias{retry_policy}
\title{The current API back-pressure policy}
\usage{
retry_policy()
}
\value{
A list with \code{max.retries}, \code{max.backoff} and
  \code{max.concurrency} (see \code{\link{set_retry_policy}})
}
\description{
The current API back-pressure policy
}
