% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_measurements.R
\name{sample_assays}
\alias{sample_assays}
\title{Which samples were measured in which measurement sets}
\usage{
sample_assays(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
\code{subject_id}, \code{sample_id}, \code{timepoint_id},
  \code{assay_name}, \code{assay_technology}, \code{measurement_set_name}
}
\description{
Which samples were measured in which measurement sets
}
