% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference.R
\name{schema_info}
\alias{schema_info}
\title{Schema name and version of a database}
\usage{
schema_info(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
List with \code{name} and \code{version}
}
\description{
Schema name and version of a database
}
