% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{select_targets}
\alias{select_targets}
\title{Keep or drop targets (columns) of a measurement matrix}
\usage{
select_targets(m, include = NULL, exclude = NULL)
}
\arguments{
\item{m}{Matrix}

\item{include, exclude}{Column names}
}
\value{
Matrix
}
\description{
Keep or drop targets (columns) of a measurement matrix
}
