% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_query_server}
\alias{set_query_server}
\title{Set the query service endpoint for this session}
\usage{
set_query_server(url)
}
\arguments{
\item{url}{The endpoint URL, e.g. \code{"https://data-commons.rcrf-dev.org"}}
}
\value{
The previous value, invisibly
}
\description{
Set the query service endpoint for this session
}
