% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_token}
\alias{set_token}
\title{Set the API token for this session}
\usage{
set_token(token)
}
\arguments{
\item{token}{The API token}
}
\value{
The previous value, invisibly
}
\description{
Overrides the \code{PATTERNQ_API_KEY} environment variable. Obtain a key
from the user settings page of the Pattern Data Commons dashboard.
}
