% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_expression.R
\name{ssgsea}
\alias{ssgsea}
\title{ssGSEA scores for several samples and gene sets}
\usage{
ssgsea(m, gene.sets = genesets(), alpha = 0.25)
}
\arguments{
\item{m}{Matrix, samples x genes (e.g. \code{to_matrix(gene_expression(...), col = "hgnc_symbol")})}

\item{gene.sets}{Named list of gene sets (default: \code{\link{genesets}()})}

\item{alpha}{Weight exponent}
}
\value{
Matrix, samples x gene sets
}
\description{
ssGSEA scores for several samples and gene sets
}
