% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_clinical.R
\name{timepoints}
\alias{timepoints}
\title{Timepoints}
\usage{
timepoints(db = NULL, ...)
}
\arguments{
\item{db}{Database name (defaults to \code{\link{current_db}})}

\item{...}{Passed to \code{\link{do_query}}}
}
\value{
\code{timepoint_id}, \code{timepoint_relative_order},
  \code{timepoint_type}, \code{timepoint_offset}, cycle/day where present,
  ordered by relative order
}
\description{
Timepoints
}
