% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{to_long}
\alias{to_long}
\title{Matrix to long format}
\usage{
to_long(m, row.name = "sample_id", col.name = "target", value.name = "value")
}
\arguments{
\item{m}{Matrix (rows = samples)}

\item{row.name, col.name, value.name}{Output column names}
}
\value{
Long data frame without \code{NA} cells
}
\description{
Matrix to long format
}
