% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_expression.R
\name{top_varying_genes}
\alias{top_varying_genes}
\title{Most variable genes}
\usage{
top_varying_genes(m, n = 500)
}
\arguments{
\item{m}{Matrix, samples x genes}

\item{n}{Number of genes}
}
\value{
\code{data.frame}: \code{hgnc_symbol}, \code{mean}, \code{variance}
}
\description{
Variance of \code{log2(1 + x)} across samples.
}
