% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{where}
\alias{where}
\title{The where portion of a query}
\usage{
where(...)
}
\arguments{
\item{...}{Clauses built with \code{\link{d}}, \code{\link{or}},
\code{\link{not}}, \code{\link{not_join}}}
}
\value{
Query section
}
\description{
The where portion of a query
}
