% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{with_args}
\alias{with_args}
\title{Add arguments to a query}
\usage{
with_args(q, ...)
}
\arguments{
\item{q}{A query}

\item{...}{Arguments appended in order}
}
\value{
The query with arguments appended
}
\description{
Add arguments to a query
}
