% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{with_vars}
\alias{with_vars}
\title{with portion of a query}
\usage{
with_vars(...)
}
\arguments{
\item{...}{Variables}
}
\value{
Query section
}
\description{
Named \code{with_vars} (datalogr called it \code{with}) so that
\code{base::with} is not masked.
}
